+++ CHANGELOG.md
@@ -24,6 +24,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [Unreleased] - 2024-01-11
+
+### REMOVED
+- GetTempCelsius method in SerialDataAcquisition.
+
+---
+
 ## [Unreleased] - 2023-12-03
 
 ### CHANGED